move(frontend): tests
